# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: api-lifecare-integrator
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:61256
  description: Generated server url
...
paths:
  /{municipalityId}/calculations:
    get:
      tags:
      - Calculations
      description: Get the calculations registered on a person in the given period
      operationId: getCalculations
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      - name: from
        in: query
        description: Start of the period
        required: false
        schema:
          type: string
          format: date
        example: 2025-01-01
      - name: to
        in: query
        description: End of the period
        required: false
        schema:
          type: string
          format: date
        example: 2026-12-31
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
        example: 1
      - name: pageSize
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
          maximum: 1000
        example: 20
      - name: ascending
        in: query
        description: Sort order
        required: false
        schema:
          type: boolean
        example: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseCalculation"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Calculations
      description: Create a calculation in Lifecare family care. Ids in the request
        come from the proposal endpoint
      operationId: createCalculation
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalculationRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedResource"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/calculations/proposal:
    get:
      tags:
      - Calculations
      description: "Get the proposal (norms, household members, income/expense types\
        \ and linkable cases) needed to create a calculation for a person. Household\
        \ members are identified by partyId — a member without a partyId could not\
        \ be resolved and cannot be referenced in a create request"
      operationId: getCalculationProposal
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculationProposal"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    CalculationExpenseRequest:
      type: object
      description: An expense or special expense included in a calculation
      properties:
        typeId:
          type: integer
          format: int32
          description: "The id of the expense type, from the calculation proposal's\
            \ expenseTypes or specialExpenseTypes"
          examples:
          - 1
        amount:
          type: number
          format: double
          description: The amount of the expense
          examples:
          - 5000.0
        approvedAmount:
          type: number
          format: double
          description: The approved amount of the expense
          examples:
          - 4500.0
        note:
          type: string
          description: A note on the expense
          examples:
          - Hyra april
      required:
      - typeId
    CalculationIncomeRequest:
      type: object
      description: An income included in a calculation
      properties:
        typeId:
          type: integer
          format: int32
          description: "The id of the income type, from the calculation proposal's\
            \ incomeTypes"
          examples:
          - 1
        applicantAmount:
          type: number
          format: double
          description: The applicant's amount
          examples:
          - 5000.0
        applicantAmountDate:
          type: string
          format: date
          description: The date of the applicant's amount
          examples:
          - 2026-05-01
        coApplicantAmount:
          type: number
          format: double
          description: The co-applicant's amount
          examples:
          - 5000.0
        coApplicantAmountDate:
          type: string
          format: date
          description: The date of the co-applicant's amount
          examples:
          - 2026-05-01
        note:
          type: string
          description: A note on the income
          examples:
          - Lön april
      required:
      - typeId
    CalculationPersonRequest:
      type: object
      description: A person included in a calculation
      properties:
        partyId:
          type: string
          description: The party id of the person
          examples:
          - 81471222-5798-11e9-ae24-57fa13b361e1
        numberOfDays:
          type: integer
          format: int32
          description: The number of days the person is included in the calculation
          examples:
          - 30
        deviationFromDate:
          type: string
          format: date
          description: The date the deviation is valid from
          examples:
          - 2026-05-01
        deviationToDate:
          type: string
          format: date
          description: The date the deviation is valid to
          examples:
          - 2026-05-31
      required:
      - partyId
    CreateCalculationRequest:
      type: object
      description: Request to create a calculation in Lifecare FC. All ids come from
        the calculation proposal endpoint.
      properties:
        partyId:
          type: string
          description: The party id of the person the calculation concerns
          examples:
          - 81471222-5798-11e9-ae24-57fa13b361e1
        serviceId:
          type: integer
          format: int32
          description: "The id of the service to link to, from the calculation proposal\
            \ endpoint"
          examples:
          - 12345
        investigationId:
          type: integer
          format: int32
          description: "The id of the investigation to link to, from the calculation\
            \ proposal endpoint"
          examples:
          - 12345
        normId:
          type: integer
          format: int32
          description: "The id of the norm to base the calculation on, from the calculation\
            \ proposal endpoint"
          examples:
          - 1
        actualisationId:
          type: integer
          format: int32
          description: "The id of the actualisation to link to, from the calculation\
            \ proposal endpoint"
          examples:
          - 12345
        calculationDate:
          type: string
          format: date
          description: The date of the calculation
          examples:
          - 2026-05-01
        calculationFromDate:
          type: string
          format: date
          description: The date the calculation is valid from
          examples:
          - 2026-05-01
        calculationToDate:
          type: string
          format: date
          description: The date the calculation is valid to
          examples:
          - 2026-05-31
        hasCustomHouseholdSize:
          type: boolean
          description: Whether a custom household size is used
          examples:
          - "false"
        householdSize:
          type: integer
          format: int32
          description: The size of the household
          examples:
          - 4
        persons:
          type: array
          description: The persons included in the calculation
          items:
            $ref: "#/components/schemas/CalculationPersonRequest"
        incomes:
          type: array
          description: The incomes included in the calculation
          items:
            $ref: "#/components/schemas/CalculationIncomeRequest"
        expenses:
          type: array
          description: The expenses included in the calculation
          items:
            $ref: "#/components/schemas/CalculationExpenseRequest"
        specialExpenses:
          type: array
          description: The special expenses included in the calculation
          items:
            $ref: "#/components/schemas/CalculationExpenseRequest"
      required:
      - calculationDate
      - calculationFromDate
      - calculationToDate
      - normId
      - partyId
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    CreatedResource:
      type: object
      description: The id of the created resource in Lifecare FC
      properties:
        id:
          type: integer
          format: int32
          description: The id of the created resource in Lifecare FC
          examples:
          - 12345
...
    Calculation:
      type: object
      description: "A calculation from Lifecare Family Care, re-modelled to a common\
        \ shape"
      properties:
        id:
          type: integer
          format: int32
          description: The calculation id in the source system
          examples:
          - 12345
        norm:
          type: string
          description: The norm the calculation is based on
          examples:
          - Riksnorm
        fromDate:
          type: string
          format: date
          description: The date the calculation is valid from
          examples:
          - 2026-05-01
        toDate:
          type: string
          format: date
          description: The date the calculation is valid to
          examples:
          - 2026-05-31
        incomeSum:
          type: number
          format: double
          description: The sum of all incomes
          examples:
          - 15000.0
        expenseSum:
          type: number
          format: double
          description: The sum of all expenses
          examples:
          - 8000.0
        specialExpenseSum:
          type: number
          format: double
          description: The sum of all special expenses
          examples:
          - 1500.0
        normSum:
          type: number
          format: double
          description: The norm sum
          examples:
          - 6320.0
        commonHouseholdCost:
          type: number
          format: double
          description: The common household cost
          examples:
          - 2000.0
        familyCost:
          type: number
          format: double
          description: The family cost
          examples:
          - 4000.0
        balance:
          type: number
          format: double
          description: The calculation balance
          examples:
          - -2820.0
        totalSum:
          type: number
          format: double
          description: The calculation total sum
          examples:
          - 2820.0
        investigationId:
          type: integer
          format: int32
          description: The id of the investigation connected to the calculation
          examples:
          - 456
        serviceId:
          type: integer
          format: int32
          description: The id of the service connected to the calculation
          examples:
          - 789
        finalCalculation:
          type: boolean
          description: Whether the calculation is final
          examples:
          - "true"
        connectedApplication:
          type: integer
          format: int32
          description: The id of the application connected to the calculation
          examples:
          - 1011
        persons:
          type: array
          description: The persons included in the calculation
          items:
            $ref: "#/components/schemas/CalculationPerson"
        incomes:
          type: array
          description: The income rows of the calculation
          items:
            $ref: "#/components/schemas/CalculationIncome"
        expenses:
          type: array
          description: The expense rows of the calculation
          items:
            $ref: "#/components/schemas/CalculationExpense"
        specialExpenses:
          type: array
          description: The special expense rows of the calculation
          items:
            $ref: "#/components/schemas/CalculationExpense"
    CalculationExpense:
      type: object
      description: "An expense row in a calculation, used for both regular and special\
        \ expenses"
      properties:
        type:
          type: string
          description: The expense type
          examples:
          - Hyra
        appliedAmount:
          type: number
          format: double
          description: The applied amount
          examples:
          - 8000.0
        approvedAmount:
          type: number
          format: double
          description: The approved amount
          examples:
          - 7500.0
    CalculationIncome:
      type: object
      description: An income row in a calculation
      properties:
        type:
          type: string
          description: The income type
          examples:
          - Lön
        amountApplicant:
          type: number
          format: double
          description: The income amount for the applicant
          examples:
          - 15000.0
        applicantSearchDate:
          type: string
          format: date
          description: The date the applicant income was searched
          examples:
          - 2026-05-01
        amountCoApplicant:
          type: number
          format: double
          description: The income amount for the co-applicant
          examples:
          - 12000.0
        coApplicantSearchDate:
          type: string
          format: date
          description: The date the co-applicant income was searched
          examples:
          - 2026-05-01
    CalculationPerson:
      type: object
      description: A person included in a calculation
      properties:
        name:
          type: string
          description: Name of the person
          examples:
          - Kalle Karlsson
        amount:
          type: number
          format: double
          description: The amount for the person
          examples:
          - 3160.0
        deviationFromDate:
          type: string
          format: date
          description: The date any deviation is valid from
          examples:
          - 2026-05-01
        deviationToDate:
          type: string
          format: date
          description: The date any deviation is valid to
          examples:
          - 2026-05-31
    PagedResponseCalculation:
      type: object
      description: "Paged response wrapper, mirroring the pagination of the underlying\
        \ Lifecare family care API"
      properties:
        page:
          type: integer
          format: int32
          description: The current page (one-based)
          examples:
          - 1
        pageSize:
          type: integer
          format: int32
          description: The number of records per page
          examples:
          - 20
        totalPages:
          type: integer
          format: int32
          description: The total number of pages
          examples:
          - 5
        totalRecords:
          type: integer
          format: int32
          description: The total number of records
          examples:
          - 98
        results:
          type: array
          description: The records on the current page
          items:
            $ref: "#/components/schemas/Calculation"
    ActualisationReference:
      type: object
      description: An existing actualisation that a calculation can be linked to
      properties:
        id:
          type: integer
          format: int32
          description: The id of the actualisation
          examples:
          - 12345
        type:
          type: string
          description: The type of the actualisation
          examples:
          - Ansökan
        date:
          type: string
          format: date
          description: The date of the actualisation
          examples:
          - 2026-05-01
    CalculationProposal:
      type: object
      description: A proposal with the valid values needed to create a calculation
        in Lifecare FC
      properties:
        investigations:
          type: array
          description: The investigations that can be linked to
          items:
            $ref: "#/components/schemas/ProposalCase"
        services:
          type: array
          description: The services that can be linked to
          items:
            $ref: "#/components/schemas/ProposalCase"
        norms:
          type: array
          description: The norms a calculation can be based on
          items:
            $ref: "#/components/schemas/Norm"
        householdMembers:
          type: array
          description: The members of the household
          items:
            $ref: "#/components/schemas/HouseholdMember"
        incomeTypes:
          type: array
          description: The valid income types
          items:
            $ref: "#/components/schemas/Lookup"
        expenseTypes:
          type: array
          description: The valid expense types
          items:
            $ref: "#/components/schemas/Lookup"
        specialExpenseTypes:
          type: array
          description: The valid special expense types
          items:
            $ref: "#/components/schemas/Lookup"
        actualisationMandatory:
          type: boolean
          description: Whether a calculation must be linked to an actualisation
          examples:
          - "true"
        numberOfFamilyMembersNotInHousehold:
          type: integer
          format: int32
          description: The number of family members not in the household
          examples:
          - 0
        actualisations:
          type: array
          description: The existing actualisations that a calculation can be linked
            to
          items:
            $ref: "#/components/schemas/ActualisationReference"
    HouseholdMember:
      type: object
      description: A member of the household
      properties:
        partyId:
          type:
          - string
          - "null"
          description: The party id of the household member; may be null when the
            member could not be resolved
          examples:
          - 81471222-5798-11e9-ae24-57fa13b361e1
        name:
          type: string
          description: The name of the household member
          examples:
          - Anna Andersson
        childFromOtherHousehold:
          type: boolean
          description: Whether the member is a child from another household
          examples:
          - "false"
    Lookup:
      type: object
      description: A lookup value from a code list
      properties:
        id:
          type: integer
          format: int32
          description: The id of the lookup value
          examples:
          - 1
        name:
          type: string
          description: The display name of the lookup value
          examples:
          - Ansökan
    Norm:
      type: object
      description: A norm that a calculation can be based on
      properties:
        id:
          type: integer
          format: int32
          description: The id of the norm
          examples:
          - 1
        name:
          type: string
          description: The name of the norm
          examples:
          - Riksnorm 2026
        validFrom:
          type: string
          format: date
          description: The date the norm is valid from
          examples:
          - 2026-01-01
        validTo:
          type: string
          format: date
          description: The date the norm is valid to
          examples:
          - 2026-12-31
    ProposalCase:
      type: object
      description: An investigation or service that can be linked to in a proposal
      properties:
        id:
          type: integer
          format: int32
          description: The id of the investigation or service
          examples:
          - 12345
        type:
          type: integer
          format: int32
          description: The type of the investigation or service
          examples:
          - 1
        name:
          type: string
          description: The name of the investigation or service
          examples:
          - Utredning ekonomiskt bistånd
        startDate:
          type: string
          format: date
          description: The start date of the investigation or service
          examples:
          - 2026-05-01
        organisationId:
          type: integer
          format: int32
          description: The id of the organization the investigation or service belongs
            to
          examples:
          - 1
        organisationUnitId:
          type: string
          description: The id of the organization unit the investigation or service
            belongs to
          examples:
          - "100"
        caseworkerId:
          type: string
          description: The id of the caseworker responsible for the investigation
            or service
          examples:
          - abc123
...
  securitySchemes: {}